Remove spurious `pytest.warns()` to fix pytest-8 compatibility
Remove the spurious `pytest.warns()` contexts within `pytest.raises()`
in `test_shortcuts`, in order to fix compatibility with pytest-8.0.0.
Prior to this version, the exception raised caused these assertions
to be ignored entirely.  This is fixed in pytest-8.0.0, and the tests
start failing because the warning is never raised prior
to the exception.

Fixes #789